Add stbt-docker: Used when video-capture hardware isn't required
Runs the stbt command in a docker container, set-up like an stb-tester ONE
but without video capture or IR control.  This is intended to be used as
the "stbt" command on MacOS and Windows allowing people to do local
development and testing (pylint, running selftests, etc.) when
video-capture is not needed.

Usage:

    $ stbt docker lint tests/tests.py
    ********** module tests
    E: 22, 4: No name '_load_template' in module 'stbt' (no-name-in-module)

    $ stbt docker exec python -m doctest tests/tests.py
    ...

    $ stbt docker exec
    Python 2.7.6 (default, Jun 22 2015, 17:58:13)
    [GCC 4.8.2] on linux2
    Type "help", "copyright", "credits" or "license" for more information.
    >>>

Environment variables:

* `DOCKER` - The location of the `docker` binary defaults to "`docker`"
* `DOCKER_OPTS` - provide additional arguments to `docker run`.  For instance
  `DOCKER_OPTS="--memory=2g"` will limit stbt to using 2GB of memory.  Shell
  escaping is used to allow specifiying arguments with spaces in.

This program is built with portability in mind so it should run on Mac OS and
Windows.  The only dependencies are Python and docker.  It is also
self-contained and relocatable so it can be deployed as a single file with no
dependency on anything else in stbt.

Note: stbt-docker hasn't actually been tested on MacOS or Windows so for
the time-being portability is an aspiration, but a realisic one.

In the future it might be worth re-writing this in Go so we can drop the
Python dependency too.