Fix stbt-docker when running on Scott's machine
It seems that docker build will extract the piped in tarball into a
temporary directory before passing the directory to the docker daemon.
This fails on Scott's linux server (Ubuntu 14.04, docker 1.8.3) with:

    unable to prepare context: unable to extract stdin to temporary context direcotry: lchown /tmp/docker-build-context-285398325/setup: operation not permitted
    Traceback (most recent call last):
      File "./stbt-docker", line 185, in <module>
        sys.exit(main(sys.argv))
      File "./stbt-docker", line 140, in main
        docker_image = build_docker_image(root)
      File "./stbt-docker", line 126, in build_docker_image
        stdout=sys.stderr, stdin=f)
      File "/usr/lib/python2.7/subprocess.py", line 540, in check_call
        raise CalledProcessError(retcode, cmd)
    subprocess.CalledProcessError: Command '['docker', 'build', '-t', 'stb-tester-test-pack-8d8f41359b26a07bdae8fff8da2502bb', '-']' returned non-zero exit status 1

because the files in the tarball has uid and gid == 0 and you are not
allowed to `chown` files to root.  For some reason this isn't a problem
on either Will's (Debian testing, docker 1.7.1) or Dave's (Ubuntu 15.04,
docker 1.6.2) linux desktop machines.